+++ src/__tests__/usePopupHistory.test.ts
@@ -1,0 +1,72 @@
+import { renderHook, act } from "@testing-library/react";
+
+import { usePopup } from "../usePopup";
+import { ICalculationsRepository } from "../data/calculations/ICalculationsRepository";
+
+const mockChrome = {
+    tabs: {
+        query: jest.fn(),
+        reload: jest.fn(),
+    },
+    runtime: {
+        onMessage: {
+            addListener: jest.fn(),
+        },
+        sendMessage: jest.fn(),
+    },
+    storage: {
+        local: {
+            get: jest.fn(),
+            set: jest.fn(),
+            onChanged: {
+                addListener: jest.fn(),
+                removeListener: jest.fn(),
+            },
+        },
+    },
+};
+
+(global as any).chrome = mockChrome;
+
+describe("usePopup", () => {
+    beforeEach(() => {
+        jest.clearAllMocks();
+    });
+
+    it("refreshCalculationHistory should show all calculations", async () => {
+        const { result } = renderHook(() => usePopup());
+
+        const mockCalculationRepository = ICalculationsRepository.instance;
+        mockCalculationRepository.storeCalculation({
+            bytes: -1,
+            emissions: -1,
+            specificEmissions: -1,
+            selectedCountries: new Map([["World Average", 0]]),
+        });
+        mockCalculationRepository.storeCalculation({
+            bytes: -2,
+            emissions: -2,
+            specificEmissions: -2,
+            selectedCountries: new Map([["World Average", 0]]),
+        });
+
+        await act(async () => {
+            await result.current.refreshCalculationHistory();
+        });
+
+        expect(result.current.calculationHistory).toStrictEqual([
+            {
+                bytes: -2,
+                emissions: -2,
+                specificEmissions: -2,
+                selectedCountries: new Map([["World Average", 0]]),
+            },
+            {
+                bytes: -1,
+                emissions: -1,
+                specificEmissions: -1,
+                selectedCountries: new Map([["World Average", 0]]),
+            },
+        ]);
+    });
+});
